commit warn system changes Lol

+++ testbot.js
@@ -5,8 +5,8 @@ const evalCmd = require('./eval.js');
 const kick = require('./kick.js');
 const ban = require('./ban.js');
 const warn = require('./warn.js');
-
-var db = require('quick.db');
+var grue = JSON.parse(fs.readFileSync("wdb.json"));
+//var db = require('quick.db');
 
 const bot = new Discord.Client({
     disableEveryone: true,
@@ -77,7 +77,7 @@ bot.on("message", async message => {
                 return;
             }
             function warnLOL(){
-                return warn(message, args, db);
+                return warn(message, args, fs, grue);
         }}
         else if(cmd === "warns") {
             if (message.author.id === info.owner){
@@ -90,11 +90,12 @@ bot.on("message", async message => {
             function cwarnLOL(){
                 const tcheckw = message.mentions.members.first();
                 if (!tcheckw){
-                    return message.channel.send("who should i check? (mention)");
+                  return message.channel.send("who should i check? (mention)");
                 }
-                let wcheck = db.get(`warnings_${message.guild.id}_${tcheckw.id}`);
-                message.channel.send(`${tcheckw} has ${wcheck} warns`);
-        }}
+                if(grue[tcheckw] == undefined) return message.channel.send(`${tcheckw} has no warns.`);
+                message.channel.send(`${tcheckw} has ${grue[tcheckw].length} warns.`);
+              }
+        }
         else if(cmd === "purge"){
             if (message.author.id === info.owner){
                 purgeLOL();
@@ -106,6 +107,15 @@ bot.on("message", async message => {
             async function purgeLOL(){
                 let grue22 = args.shift();
                 let aelol = Number(grue22) + 1;
+                if (aelol >= 101){
+                    return message.channel.send("can't purge more than 100 msgs yo");
+                } if (aelol == undefined || aelol == -Infinity || aelol == NaN){
+                    return message.channel.send("what the fuck is wrong with you, specify a simple number goddamit");
+                }
+                if (aelol <= 0) {
+                    return message.channel.send("how many messages should i purge? (max 100)");
+                } 
+                
                 console.log(aelol);
                 await message.channel.bulkDelete(aelol);
         }}
